Add Computed Propertiers
Adding attributes to Order Item to indicate the product related to one order item and the total of products per item.

Creating a computed property that calculates the Item's Total Price =  Product's Price + Quantity.

Creating computed properties in Order to calc the Total Quantity of Products and the Grand Total Price.